+++ src/wiktextract/extractor/en/pronunciation.py
@@ -4,12 +4,13 @@ import urllib
 from typing import Iterator, Optional, Union
 
 from wikitextprocessor import NodeKind, WikiNode
+from wikitextprocessor.parser import TemplateNode
 from wiktextract.datautils import data_append, split_at_comma_semi
 from wiktextract.form_descriptions import (
     classify_desc,
     parse_pronunciation_tags,
 )
-from wiktextract.page import clean_node, is_panel_template
+from wiktextract.page import LEVEL_KINDS, clean_node, is_panel_template
 from wiktextract.parts_of_speech import part_of_speech_map
 from wiktextract.tags import valid_tags
 from wiktextract.type_utils import SoundData, TemplateArgs, WordData
@@ -17,20 +18,11 @@ from wiktextract.wxr_context import WiktextractContext
 
 from .zh_pron_tags import ZH_PRON_TAGS
 
-LEVEL_KINDS = (
-    NodeKind.LEVEL2,
-    NodeKind.LEVEL3,
-    NodeKind.LEVEL4,
-    NodeKind.LEVEL5,
-    NodeKind.LEVEL6,
-)
-
 # Prefixes, tags, and regexp for finding romanizations from the pronuncation
 # section
 pron_romanizations = {
     " Revised Romanization ": "romanization revised",
-    " Revised Romanization (translit.) ":
-        "romanization revised transliteration",
+    " Revised Romanization (translit.) ": "romanization revised transliteration",
     " McCune-Reischauer ": "McCune-Reischauer romanization",
     " McCune–Reischauer ": "McCune-Reischauer romanization",
     " Yale Romanization ": "Yale romanization",
@@ -64,14 +56,14 @@ def parse_pronunciation(
     # Remove subsections, such as Usage notes.  They may contain IPAchar
     # templates in running text, and we do not want to extract IPAs from
     # those.
+    # Filter out only LEVEL_KINDS; 'or' is doing heavy lifting here
+    # Slip through not-WikiNodes, then slip through WikiNodes that
+    # are not LEVEL_KINDS.
     contents = [
         x
         for x in contents
         if not isinstance(x, WikiNode) or x.kind not in LEVEL_KINDS
     ]
-    # Filter out only LEVEL_KINDS; 'or' is doing heavy lifting here
-    # Slip through not-WikiNodes, then slip through WikiNodes that
-    # are not LEVEL_KINDS.
     if not any(
         isinstance(x, WikiNode) and x.kind == NodeKind.LIST for x in contents
     ):
@@ -79,8 +71,7 @@ def parse_pronunciation(
         new_contents: list[Union[str, WikiNode]] = []
         for lst in contents:
             if (
-                isinstance(lst, WikiNode)
-                and lst.kind == NodeKind.TEMPLATE
+                isinstance(lst, TemplateNode)
                 and isinstance(lst.largs[0][0], str)
                 and lst.largs[0][0].strip() != "zh-pron"
             ):
@@ -497,7 +488,7 @@ def parse_pronunciation(
     active_pos: Optional[str] = None
 
     for text, ipa_text in split_cleaned_node_on_newlines(contents):
-        # print(f"{text=}, {ipa_text=}")
+        print(f"{text=}, {ipa_text=}")
         prefix: Optional[str] = None
         if not text:
             continue
@@ -523,8 +514,8 @@ def parse_pronunciation(
 
         m = re.match(r"\s*(\w+\s?\w*)", text)
         if m:
-            if m.group(1).lower() in part_of_speech_map:
-                active_pos = part_of_speech_map[m.group(1).lower()]["pos"]
+            if (m_lower := m.group(1).lower()) in part_of_speech_map:
+                active_pos = part_of_speech_map[m_lower]["pos"]
 
         if "IPA" in text:
             field = "ipa"
@@ -656,77 +647,77 @@ def parse_pronunciation(
         # XXX what about {{hyphenation|...}}, {{hyph|...}}
         # and those used to be stored under "hyphenation"
 
-    # Add data that was collected in template_fn
-    for audio in audios:
-        if "audio" in audio:
-            # Compute audio file URLs
-            fn = audio["audio"]
-            # Strip certain characters, e.g., left-to-right mark
-            fn = re.sub(r"[\u200f\u200e]", "", fn)
-            fn = fn.strip()
-            fn = urllib.parse.unquote(fn)
-            # First character is usually uppercased
-            if re.match(r"^[a-z][a-z]+", fn):
-                fn = fn[0].upper() + fn[1:]
-            if fn in wxr.config.redirects:
-                fn = wxr.config.redirects[fn]
-            # File extension is lowercased
-            # XXX some words seem to need this, some don't seem to
-            # have this??? what is the exact rule?
-            # fn = re.sub(r"\.[^.]*$", lambda m: m.group(0).lower(), fn)
-            # Spaces are converted to underscores
-            fn = re.sub(r"\s+", "_", fn)
-            # Compute hash digest part
-            h = hashlib.md5()
-            hname = fn.encode("utf-8")
-            h.update(hname)
-            digest = h.hexdigest()
-            # Quote filename for URL
-            qfn = urllib.parse.quote(fn)
-            # For safety when writing files
-            qfn = qfn.replace("/", "__slash__")
-            if re.search(r"(?i)\.(ogg|oga)$", fn):
-                ogg = (
-                    "https://upload.wikimedia.org/wikipedia/"
-                    "commons/{}/{}/{}".format(digest[:1], digest[:2], qfn)
-                )
-            else:
-                ogg = (
-                    "https://upload.wikimedia.org/wikipedia/"
-                    "commons/transcoded/"
-                    "{}/{}/{}/{}.ogg".format(digest[:1], digest[:2], qfn, qfn)
-                )
-            if re.search(r"(?i)\.(mp3)$", fn):
-                mp3 = (
-                    "https://upload.wikimedia.org/wikipedia/"
-                    "commons/{}/{}/{}".format(digest[:1], digest[:2], qfn)
-                )
-            else:
-                mp3 = (
-                    "https://upload.wikimedia.org/wikipedia/"
-                    "commons/transcoded/"
-                    "{}/{}/{}/{}.mp3".format(digest[:1], digest[:2], qfn, qfn)
-                )
-            audio["ogg_url"] = ogg
-            audio["mp3_url"] = mp3
+        # Add data that was collected in template_fn
+        for audio in audios:
+            if "audio" in audio:
+                # Compute audio file URLs
+                fn = audio["audio"]
+                # Strip certain characters, e.g., left-to-right mark
+                fn = re.sub(r"[\u200f\u200e]", "", fn)
+                fn = fn.strip()
+                fn = urllib.parse.unquote(fn)
+                # First character is usually uppercased
+                if re.match(r"^[a-z][a-z]+", fn):
+                    fn = fn[0].upper() + fn[1:]
+                if fn in wxr.config.redirects:
+                    fn = wxr.config.redirects[fn]
+                # File extension is lowercased
+                # XXX some words seem to need this, some don't seem to
+                # have this??? what is the exact rule?
+                # fn = re.sub(r"\.[^.]*$", lambda m: m.group(0).lower(), fn)
+                # Spaces are converted to underscores
+                fn = re.sub(r"\s+", "_", fn)
+                # Compute hash digest part
+                h = hashlib.md5()
+                hname = fn.encode("utf-8")
+                h.update(hname)
+                digest = h.hexdigest()
+                # Quote filename for URL
+                qfn = urllib.parse.quote(fn)
+                # For safety when writing files
+                qfn = qfn.replace("/", "__slash__")
+                if re.search(r"(?i)\.(ogg|oga)$", fn):
+                    ogg = (
+                        "https://upload.wikimedia.org/wikipedia/"
+                        "commons/{}/{}/{}".format(digest[:1], digest[:2], qfn)
+                    )
+                else:
+                    ogg = (
+                        "https://upload.wikimedia.org/wikipedia/"
+                        "commons/transcoded/"
+                        "{}/{}/{}/{}.ogg".format(digest[:1], digest[:2], qfn, qfn)
+                    )
+                if re.search(r"(?i)\.(mp3)$", fn):
+                    mp3 = (
+                        "https://upload.wikimedia.org/wikipedia/"
+                        "commons/{}/{}/{}".format(digest[:1], digest[:2], qfn)
+                    )
+                else:
+                    mp3 = (
+                        "https://upload.wikimedia.org/wikipedia/"
+                        "commons/transcoded/"
+                        "{}/{}/{}/{}.mp3".format(digest[:1], digest[:2], qfn, qfn)
+                    )
+                audio["ogg_url"] = ogg
+                audio["mp3_url"] = mp3
+                if active_pos:
+                    audio["pos"] = active_pos  # type: ignore[typeddict-unknown-key]
+            if audio not in data.get("sounds", ()):
+                data_append(data, "sounds", audio)
+        # if audios:
+        #     have_pronunciations = True
+        audios = []
+        for enpr in enprs:
+            if re.match(r"/[^/]+/$", enpr):
+                enpr = enpr[1:-1]
+            pron = {"enpr": enpr}
+            parse_pronunciation_tags(wxr, tagstext, pron)
             if active_pos:
-                audio["pos"] = active_pos  # type: ignore[typeddict-unknown-key]
-        if audio not in data.get("sounds", ()):
-            data_append(data, "sounds", audio)
-    # if audios:
-    #     have_pronunciations = True
-    audios = []
-    for enpr in enprs:
-        if re.match(r"/[^/]+/$", enpr):
-            enpr = enpr[1:-1]
-        pron = {"enpr": enpr}
-        parse_pronunciation_tags(wxr, tagstext, pron)
-        if active_pos:
-            pron["pos"] = active_pos  # type: ignore[typeddict-unknown-key]
-        if pron not in data.get("sounds", ()):
-            data_append(data, "sounds", pron)
-        # have_pronunciations = True
-    enprs = []
+                pron["pos"] = active_pos  # type: ignore[typeddict-unknown-key]
+            if pron not in data.get("sounds", ()):
+                data_append(data, "sounds", pron)
+            # have_pronunciations = True
+        enprs = []
 
     ## I have commented out the otherwise unused have_pronunciation
     ## toggles; uncomment them to use this debug print
